melhorias, adição tratamento de excessão, metodo put, documentação dom swagger, mensagens de erro em um arquivo .properties

+++ src/main/java/com/github/aniversario/service/BirthdayPersonService.java
@@ -1,11 +1,14 @@
 package com.github.aniversario.service;
 
-import com.github.aniversario.dto.BirthdayPersonDTO;
+import com.github.aniversario.dto.BirthdayPersonDto;
 import com.github.aniversario.entities.BirthdayPerson;
 import com.github.aniversario.repositories.BirthdayPersonRepository;
+import exception.BirthdayPersonNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -14,26 +17,59 @@ public class BirthdayPersonService {
     @Autowired
     private BirthdayPersonRepository birthdayPersonRepository;
 
-    public List<BirthdayPersonDTO> findAll() {
+    public static final String MESSAGE_EXCEPTION_NOT_FOUND = "Aniversariante não encontrado";
+
+    public List<BirthdayPersonDto> findAll() {
         List<BirthdayPerson> result = birthdayPersonRepository.findAll();
-        return result.stream().map(x -> new BirthdayPersonDTO(x)).collect(Collectors.toList());
+        return result.stream().map(x -> new BirthdayPersonDto(x)).collect(Collectors.toList());
     }
 
-    public List<BirthdayPersonDTO> birthdayOfTheDay(Integer day, Integer month) {
+    public BirthdayPersonDto findById(Long id) {
+        Optional<BirthdayPerson> birthdayPerson = birthdayPersonRepository.findById(id);
+        if (birthdayPerson.isPresent())
+            return new BirthdayPersonDto(birthdayPerson.get());
+        throw new BirthdayPersonNotFoundException(MESSAGE_EXCEPTION_NOT_FOUND);
+    }
+
+    public List<BirthdayPersonDto> findByAllName(String name) {
+        List<BirthdayPerson> result = birthdayPersonRepository.findAllByName(name);
+        return result.stream().map(x -> new BirthdayPersonDto(x)).collect(Collectors.toList());
+    }
+
+    public List<BirthdayPersonDto> findByAllDescription(String description) {
+        List<BirthdayPerson> result = birthdayPersonRepository.findAllByDescription(description);
+        return result.stream().map(x -> new BirthdayPersonDto(x)).collect(Collectors.toList());
+    }
+
+    public List<BirthdayPersonDto> findByAllDayAndMonth(Integer day, Integer month) {
         List<BirthdayPerson> result = birthdayPersonRepository.findByDayAndMonth(day, month);
-        return result.stream().map(x -> new BirthdayPersonDTO(x)).collect(Collectors.toList());
+        return result.stream().map(x -> new BirthdayPersonDto(x)).collect(Collectors.toList());
     }
 
-    public List<BirthdayPersonDTO> birthOfTheMonth(Integer month) {
+    public List<BirthdayPersonDto> findByAllMonth(Integer month) {
         List<BirthdayPerson> result = birthdayPersonRepository.findByMonth(month);
-        return result.stream().map(x -> new BirthdayPersonDTO(x)).collect(Collectors.toList());
+        return result.stream().map(x -> new BirthdayPersonDto(x)).collect(Collectors.toList());
     }
 
-    public BirthdayPerson save(BirthdayPerson birthdayPerson) {
-        return birthdayPersonRepository.save(birthdayPerson);
+    public BirthdayPersonDto save(BirthdayPersonDto birthdayPersonDto) {
+        BirthdayPerson birthdayPerson = birthdayPersonRepository.save(new BirthdayPerson(birthdayPersonDto));
+        return new BirthdayPersonDto(birthdayPerson);
+    }
+
+    public BirthdayPersonDto update(BirthdayPersonDto birthdayPersonDto) {
+        Optional<BirthdayPerson> birthdayPerson = birthdayPersonRepository.findById(birthdayPersonDto.getId());
+        if (birthdayPerson.isPresent()) {
+            birthdayPerson.get().update(birthdayPersonDto);
+            return new BirthdayPersonDto(birthdayPersonRepository.save(birthdayPerson.get()));
+        }
+        throw new BirthdayPersonNotFoundException(MESSAGE_EXCEPTION_NOT_FOUND);
     }
 
     public void delete(long id) {
-        birthdayPersonRepository.deleteById(id);
+        Optional<BirthdayPerson> birthdayPerson = birthdayPersonRepository.findById(id);
+        if (birthdayPerson.isPresent())
+            birthdayPersonRepository.deleteById(id);
+        else
+            throw new BirthdayPersonNotFoundException(MESSAGE_EXCEPTION_NOT_FOUND);
     }
 }
